+++ pkg/core/policy/policy_suite_test.go
@@ -1,0 +1,13 @@
+package policy_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestMesh(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Policy Suite")
+}